LayoutTests/fast/regions/webkit-named-flow-get-regions-pseudo.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/helper.js"></script>
<style>
    #article {
        -webkit-flow-into: article;
        height: 50px;
        width: 50px;
        background-color: green;
    }

    #region_1 {
        position: absolute;
        top:200px;
        left: 100px;
    }

    #region_1::before {
        display: block;
        width: 50px;
        height: 50px;
        border: 1px solid black;
        -webkit-flow-from: article;
    }

    #region_1::after {
        display: block;
        width: 50px;
        height: 50px;
        border: 1px solid black;
        -webkit-flow-from: article;    
    }

    #region_2 {
        -webkit-flow-from: article;
        height:50px;
        width:50px;
        position: absolute;
        top: 200px;
        left: 200px;
        border: 1px solid black;
    }

</style>
</head>
    <body>
        <!--
        Test that getRegions() function from NamedFlow interface does not expose the pseudo-elements as
        regions.
        -->
        <div id="article"></div>

        <div id="region_1" class="region_1"></div>
        <div id="region_2" class="region_2"></div>

        <script>
            description("Tests WebKitNamedFlow.getRegions() and pseudo-elements as regions.")
            if (window.testRunner)
                testRunner.dumpAsText();

            var namedFlow = getFlowByName("article");

            // The pseudo-elements are not yet returned by getRegions()
            var regionList = namedFlow.getRegions();
            shouldBe("regionList.length", "1");
            shouldBe("namedFlow.firstEmptyRegionIndex", "0");
            shouldBeEqualToString("regionList.item(0).id", "region_2");
            shouldBeEqualToString("regionList.item(0).webkitRegionOverset", "empty");

            document.getElementById("article").style.visibility = "hidden";
            document.getElementById("region_1").style.visibility = "hidden";
            document.getElementById("region_2").style.visibility = "hidden";
        </script>
        <script src="../../fast/js/resources/js-test-post.js"></script>
    </body>
</html>
