LayoutTests/fast/regions/webkit-named-flow-get-regions-pseudo-expected.txt

Tests WebKitNamedFlow.getRegions() and pseudo-elements as regions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regionList.length is 1
PASS namedFlow.firstEmptyRegionIndex is 0
PASS regionList.item(0).id is "region_2"
PASS regionList.item(0).webkitRegionOverset is "empty"
PASS successfullyParsed is true

TEST COMPLETE
